requirements.txt
fastapi
uvicorn[standard]
motor
beanie
python-jose[cryptography]
passlib[bcrypt]
python-multipart
web3
requests
qrcode[pil]
Pillow
transformers
torch
python-dotenv
pydantic-settings
email-validator
